Backend/API/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
openai>=1.0.0
python-dotenv
flask-cors
gunicorn
requests
pytest
flask_testing
supabase>=1.0.3
python-jose[cryptography]>=3.3.0
postgrest>=0.10.6
